Add tests for macros
