+++ test/cfg/macros/alignak_macroresolver.cfg
@@ -1,0 +1,24 @@
+cfg_dir=../default
+
+$USER1$=plugins
+$INTERESTINGVARIABLE$=interesting_value
+$ANOTHERVALUE$=first=second
+
+define service{
+    active_checks_enabled       1
+    check_command               check_service!ok
+    check_interval              1
+    host_name                   test_host_0
+    icon_image                  ../../docs/images/tip.gif?host=$HOSTNAME$&srv=$SERVICEDESC$
+    icon_image_alt              icon alt string
+    notes                       just a notes string
+    retry_interval              1
+    service_description         test_another_service
+    servicegroups               servicegroup_01,ok
+    use                         generic-service
+    event_handler               eventhandler
+    notes_url                   /alignak/wiki/doku.php/$HOSTNAME$/$SERVICEDESC$
+    action_url                  /alignak/pnp/index.php?host=$HOSTNAME$&srv=$SERVICEDESC$
+    _custom1			        value
+    _custom2			        $HOSTNAME$
+}
